requirements.txt
numpy==1.13.1
Pillow==4.2.1
lmdb==0.93
